%$Date$
%$Author$
\name{Simple Decay Example}
\alias{DecayData}
\alias{DecayModel}
\alias{DecayMod}

\docType{data}

\title{A simple decay model}

\description{
This datasets contains a simple decay model which has a single species X and a single degradation reaction with rate mu per molecule of X.}
\usage{data(Decay)}
  
\format{

    \code{DecayData}: A simulated stochastic data set, using the standard Gillespie algorithm.
    
    \code{DecayModel}: A SBML string representing a simple decay (or death) process.
    
    \code{DecayMod}: The SBML shorthand representation of the model.

    }
        
\keyword{datasets}
