%$Date$
%$Author$
\name{SBMLModels-package}
\alias{SBMLModels-package}
\alias{SBMLModels}


\docType{package}

\title{A collection of SBML Models}

\description{ A collection of SBML Models with experimental and/or simulated datasets. Each data set contains a SBML model, a SBML shorthand model and associated datasets.

SBML-shorthand is a language for describing systems biology models. It is designed
primarily for translation (compilation) into SBML. It is not associated with any particular
modelling or simulation tool. SBML itself is not intended to be read by humans or written
hand -- SBML-shorthand is. Note that SBML-shorthand is not intended to cover
every feature in SBML -- just the most commonly used aspects.

SBML-shorthand files have the file extension .mod.


}

\author{
Maintainer: Colin Gillespie <c.gillespie@ncl.ac.uk>
}
\references{
SBML \url{http://www.sbml.org/}

Wilkinson, D. J., \url{http://www.staff.ncl.ac.uk/d.j.wilkinson/software/sbml-sh/}
}


\keyword{package}
\keyword{distribution}
